+++ HISTORY.md
@@ -4,6 +4,8 @@ Express YUI Change History
 @NEXT@
 ------------------
 
+* default lang bundle should be used if a custom lang bundle is not available
+
 1.1.1 (2014-02-11)
 ------------------